Revert "UPSTREAM: usb: dwc3: gadget: cope with XferNotReady before usb_ep_queue()"
This reverts commit c139ec131a38532a29708319e296cac8eb56562c.

This patch needs a series of related patches, if we only pick
this upstream patch, the UVC gadget can't work normally. So
just revert it.

As a side effect, the audio source function doesn't work.
However, the audio source function is a only used for linux
linaro, and we suggest to use UAC1 or UAC2 instead of audio source.

Change-Id: I82a69f122dbf48eda63893405a67d35e6d9cad80
Signed-off-by: William Wu <william.wu@rock-chips.com>

rockchip 54bf73cae73da3e758ae91bd4e7afe974927b321