esp8089: Fix 2nd probe issue after firmware load via sdio
ESP8089 sdio device has a multiple stage bring-up process.
it loads firmware on the first call to its sdio-drivers' probe
and then resets the device causeing it to reboot from its RAM
with the new firmware.

This patch fixes a rescan problem caused during this process.

This commit is from the following patches,
and for esp8089 driver of odroid-goA, a part of related patch is
applied according to current esp8089 driver version.

http://patches.linaro.org/project/linux-kernel/list/?series=2972
1. esp8089 driver
http://patches.linaro.org/patch/108483/
2. mmc core driver
https://patchwork.kernel.org/patch/9857213/yy

Change-Id: Ieee29e01d4b69f3fd137ae5624ba629524794374