ODROID-XU4: dts/overlays: Change the SPI frequency of HK ILI9340 based LCDs
The small LCD using the ILI9340 controller can run at 40 MHz SPI clock
speed. but in XU4, since the XU4's SPI could be unstable at high
frequency, it may occur an unexpected error at 40 MHz so that the LCD
may not work at all except the backlight over time. So it is needed to
change the SPI max frequency to 32 MHz for XU4's SPI capability. This
barely affects the performance.

Signed-off-by: Deokgyu Yang <secugyu@gmail.com>
Change-Id: I464367fdc147affd26ca207941ad581e7b181e80