power: supply: sc8886: stop probe when battery was no found
SC8886 will shutdown the output off in these cases
when battery was removed:
1. INPUT CRRENT is set to 0. In Battery system, battery
would take over system supply, but in no battery system,
it would restart due to system poweroff.

2. VBat is higher than VSysmin, sc8886 would switch
system supply to VBat, But in no battery system, there
is just a capacitance, the voltage of capacitance would
be pull down quickly when supply switch and it make
system restart.

Signed-off-by: Zain Wang <wzz@rock-chips.com>
Change-Id: I0398d55d511254821744c4e7d4448ef4605af509